+++ tests/test_gitwildmatch.py
@@ -774,6 +774,18 @@ class GitWildMatchTest(unittest.TestCase):
 			'anydir/file.txt',
 		})
 
+	def test_13_issue_77_regex(self):
+		"""
+		Test the resulting regex for regex bracket expression negation.
+		"""
+		regex, include = GitWildMatchPattern.pattern_to_regex('a[^b]c')
+		self.assertTrue(include)
+
+		equiv_regex, include = GitWildMatchPattern.pattern_to_regex('a[!b]c')
+		self.assertTrue(include)
+
+		self.assertEqual(regex, equiv_regex)
+
 	def test_13_negate_with_caret(self):
 		"""
 		Test negation using the caret symbol (^)