+++ CHANGES.rst
@@ -3,6 +3,19 @@ Change History
 ==============
 
 
+0.11.2 (TBD)
+------------
+
+Bug fixes:
+
+- `Pull #76`_: Add edge case: patterns that end with an escaped space
+- `Issue #77`_: On bracket expression negation.
+
+
+.. _`Pull #76`: https://github.com/cpburnz/python-pathspec/pull/76
+.. _`Issue #77`: https://github.com/cpburnz/python-pathspec/issues/77
+
+
 0.11.1 (2023-03-14)
 -------------------
 
@@ -15,9 +28,9 @@ Improvements:
 - `Pull #75`_: Fix partially unknown PathLike type.
 - Convert `os.PathLike` to a string properly using `os.fspath`.
 
+
 .. _`Issue #74`: https://github.com/cpburnz/python-pathspec/issues/74
 .. _`Pull #75`: https://github.com/cpburnz/python-pathspec/pull/75
-
 
 
 0.11.0 (2023-01-24)